Add resources to player resources and separation of concerns
