+++ src/Troop.cpp
@@ -1,0 +1,16 @@
+// Troop.cpp
+#include "Troop.h"
+
+std::vector<Troop> playerTroops = {
+    {"Porter", 1, 10, 45, 100, 200, 2, 1, 0},
+    {"Conscript", 10, 10, 75, 200, 20, 3, 1, 0},
+    {"Spy", 5, 5, 10, 3000, 0, 5, 2, 0},
+    {"Halberdsman", 40, 40, 150, 300, 40, 6, 2, 0},
+    {"Minotaur", 70, 45, 225, 275, 30, 7, 3, 0},
+    {"Longbowman", 5, 30, 75, 250, 25, 9, 4, 0},
+    {"Swift Strike Dragon", 150, 60, 300, 1000, 100, 18, 5, 0},
+    {"Battle Dragon", 300, 300, 1500, 750, 80, 35, 7, 0},
+    {"Armored Transport", 5, 200, 750, 150, 5000, 10, 6, 0},
+    {"Giant", 1000, 400, 4000, 120, 45, 100, 9, 0},
+    {"Fire Mirror", 20, 30, 1500, 50, 75, 250, 10, 0},
+    {"Pack Dragon", 150, 400, 850, 1000, 6000, 10, 6, 0}};
